requirements.txt
bidict==0.23.1
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==2.1.1
click==8.1.7
Flask==3.0.3
Flask-SocketIO==5.3.6
h11==0.14.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.0
python-engineio==4.9.1
python-socketio==5.11.3
requests==2.28.1
simple-websocket==1.0.0
urllib3==1.26.19
Werkzeug==3.0.4
wsproto==1.2.0
